# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Office Task Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - name: category_id
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: created_by
          in: query
          schema:
            type: string
        - name: is_archived
          in: query
          schema:
            type: boolean
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [all, my_tasks, created_by_me, delegated]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Task deleted
  /tasks/{id}/archive:
    post:
      operationId: archiveTask
      tags: [tasks]
      summary: Archive or restore a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_archived:
                  type: boolean
              required:
                - is_archived
      responses:
        "200":
          description: Task archive status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/status:
    post:
      operationId: changeTaskStatus
      tags: [tasks]
      summary: Change task status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/TaskStatus"
              required:
                - status
      responses:
        "200":
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

    post:
      operationId: createUser
      tags: [users]
      summary: Create/invite a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/active:
    get:
      operationId: getActiveUsers
      tags: [users]
      summary: List active users (minimal) for task assignment — available to all authenticated users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of active users with minimal info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDirectory"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    UserDirectory:
      type: object
      description: Minimal user info for task assignment dropdowns
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        department:
          type: string
          nullable: true
      required:
        - id
        - email
        - name
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        department:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - email
        - name
        - role
        - is_active
        - created_at
        - updated_at
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
        department:
          type: string
      required:
        - email
        - name
        - password
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        department:
          type: string
        is_active:
          type: boolean
        avatar_url:
          type: string
    TaskStatus:
      type: string
      enum:
        - not_started
        - in_progress
        - waiting_for_response
        - deferred
        - completed
        - cancelled
    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - urgent
    ReminderOption:
      type: string
      enum:
        - none
        - on_due
        - 15min_before
        - 1hr_before
        - 2hr_before
        - 1day_before
        - custom
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category_id:
          type: string
          nullable: true
        reference_number:
          type: string
          nullable: true
        source_department:
          type: string
          nullable: true
        assigned_to:
          type: string
          nullable: true
        assigned_to_name:
          type: string
          nullable: true
        created_by:
          type: string
        created_by_name:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        start_date:
          type: string
          nullable: true
        due_date:
          type: string
        due_time:
          type: string
          nullable: true
        reminder_option:
          $ref: "#/components/schemas/ReminderOption"
        custom_reminder_datetime:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
        is_archived:
          type: boolean
        completed_at:
          type: string
          nullable: true
        reminder_sent:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - title
        - created_by
        - created_by_name
        - priority
        - status
        - due_date
        - reminder_option
        - is_archived
        - reminder_sent
        - created_at
        - updated_at
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
        reference_number:
          type: string
        source_department:
          type: string
        assigned_to:
          type: string
        assigned_to_name:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        start_date:
          type: string
        due_date:
          type: string
        due_time:
          type: string
        reminder_option:
          $ref: "#/components/schemas/ReminderOption"
        custom_reminder_datetime:
          type: string
        tags:
          type: string
      required:
        - title
        - due_date
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
        reference_number:
          type: string
        source_department:
          type: string
        assigned_to:
          type: string
        assigned_to_name:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        start_date:
          type: string
        due_date:
          type: string
        due_time:
          type: string
        reminder_option:
          $ref: "#/components/schemas/ReminderOption"
        custom_reminder_datetime:
          type: string
        tags:
          type: string
...