# lib/api-zod/src/generated/types/userDirectory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office Task Management System API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Minimal user info for task assignment dropdowns
 */
export interface UserDirectory {
  id: string;
  email: string;
  name: string;
  department?: string | null;
}